modifié l'architecture de Strategy et StrategyBook commencé un sélecteur simple de stratégie en fonction de la position de la balle
